cdk/requirements.txt
aws-cdk.core==1.116.0
aws-cdk.aws-iam==1.116.0
aws-cdk.aws-sagemaker==1.116.0
aws-cdk.aws-stepfunctions==1.116.0 
aws-cdk.aws-stepfunctions-tasks==1.116.0 
aws-cdk.aws-sns==1.116.0
aws-cdk.aws-sns-subscriptions==1.116.0
aws-cdk.aws-s3==1.116.0
